+++ src/pages/task/newTask/IndexNewTask.tsx
@@ -46,8 +46,8 @@ import { color, icon } from "../../../resource";
 import { DateTimeUtil } from "../../../utilities/DateTimeUtil";
 import { numberWithCommas } from "../../../utilities/TextFormatter";
 import { useNavigate } from "react-router-dom";
-import { ListCheckHaveLine } from "../../../components/dropdownCheck/ListStatusFarmer";
 import { listAppType } from "../../../definitions/ApplicatoionTypes";
+import { ListCheckHaveLine } from "../../../components/dropdownCheck/ListStatusAppType";
 
 const { RangePicker } = DatePicker;
 const dateFormat = "DD-MM-YYYY";