+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Version - 0]
 
+## improve [v0.36.8] 2023-08-15
+
+- improve index droner page
+
 ## Fix [v0.36.7] 2023-08-11
 
 - fixed validate form farmer plot